+++ test/specs/graphql/test_basic.py
@@ -15,7 +15,13 @@ from schemathesis.extra._flask import run_server as run_flask_server
 from schemathesis.specs.graphql.loaders import extract_schema_from_response, get_introspection_query
 from schemathesis.specs.graphql.schemas import GraphQLCase
 from schemathesis.specs.graphql.validation import validate_graphql_response
-from schemathesis.specs.openapi.checks import ensure_resource_availability, ignored_auth, use_after_free
+from schemathesis.specs.openapi.checks import (
+    ensure_resource_availability,
+    ignored_auth,
+    negative_data_rejection,
+    positive_data_acceptance,
+    use_after_free,
+)
 from schemathesis.transports import WSGITransport
 from test.apps import _graphql as graphql
 from test.apps._graphql.schema import Author
@@ -349,7 +355,10 @@ def test_schema_as_strategy(graphql_schema):
             find(strategy, lambda x, op=operation: op.definition.field_name in x.body)
 
 
-@pytest.mark.parametrize("check", (use_after_free, ensure_resource_availability, ignored_auth))
+@pytest.mark.parametrize(
+    "check",
+    (use_after_free, ensure_resource_availability, ignored_auth, positive_data_acceptance, negative_data_rejection),
+)
 def test_ignored_checks(graphql_schema, check):
     # Just in case
     case = graphql_schema["Query"]["getBooks"].make_case()